+++ core/impl/src/main/java/com/blazebit/persistence/impl/plan/CustomSelectQueryPlan.java
@@ -31,6 +31,7 @@ public class CustomSelectQueryPlan<T> implements SelectQueryPlan<T> {
 
     private final ExtendedQuerySupport extendedQuerySupport;
     private final ServiceProvider serviceProvider;
+    private final Query baseQuery;
     private final Query delegate;
     private final List<Query> participatingQueries;
     private final String sql;
@@ -38,9 +39,10 @@ public class CustomSelectQueryPlan<T> implements SelectQueryPlan<T> {
     private final int maxResults;
     private final boolean queryPlanCacheEnabled;
 
-    public CustomSelectQueryPlan(ExtendedQuerySupport extendedQuerySupport, ServiceProvider serviceProvider, Query delegate, List<Query> participatingQueries, String sql, int firstResult, int maxResults, boolean queryPlanCacheEnabled) {
+    public CustomSelectQueryPlan(ExtendedQuerySupport extendedQuerySupport, ServiceProvider serviceProvider, Query baseQuery, Query delegate, List<Query> participatingQueries, String sql, int firstResult, int maxResults, boolean queryPlanCacheEnabled) {
         this.extendedQuerySupport = extendedQuerySupport;
         this.serviceProvider = serviceProvider;
+        this.baseQuery = baseQuery;
         this.delegate = delegate;
         this.participatingQueries = participatingQueries;
         this.sql = sql;
@@ -51,15 +53,15 @@ public class CustomSelectQueryPlan<T> implements SelectQueryPlan<T> {
 
     @Override
     public List<T> getResultList() {
-        delegate.setFirstResult(firstResult);
-        delegate.setMaxResults(maxResults);
+        baseQuery.setFirstResult(firstResult);
+        baseQuery.setMaxResults(maxResults);
         return extendedQuerySupport.getResultList(serviceProvider, participatingQueries, delegate, sql, queryPlanCacheEnabled);
     }
 
     @Override
     public T getSingleResult() {
-        delegate.setFirstResult(firstResult);
-        delegate.setMaxResults(maxResults);
+        baseQuery.setFirstResult(firstResult);
+        baseQuery.setMaxResults(maxResults);
         return (T) extendedQuerySupport.getSingleResult(serviceProvider, participatingQueries, delegate, sql, queryPlanCacheEnabled);
     }
 }