+++ core/impl/src/main/java/com/blazebit/persistence/impl/query/SetOperationQuerySpecification.java
@@ -47,7 +47,7 @@ public class SetOperationQuerySpecification<T> extends CustomQuerySpecification<
                                           List<? extends OrderByElement> orderByElements, boolean nested, Set<Parameter<?>> parameters, Set<String> parameterListNames, String limit, String offset,
                                           List<String> keyRestrictedLeftJoinAliases, List<EntityFunctionNode> entityFunctionNodes, boolean recursive, List<CTENode> ctes, boolean shouldRenderCteNodes,
                                           boolean queryPlanCacheEnabled) {
-        super(commonQueryBuilder, baseQuery, parameters, parameterListNames, limit, offset, keyRestrictedLeftJoinAliases, entityFunctionNodes, recursive, ctes, shouldRenderCteNodes, queryPlanCacheEnabled);
+        super(commonQueryBuilder, baseQuery, parameters, parameterListNames, limit, offset, keyRestrictedLeftJoinAliases, entityFunctionNodes, recursive, ctes, shouldRenderCteNodes, queryPlanCacheEnabled, null);
         this.leftMostQuery = leftMostQuery;
         this.setOperands = setOperands;
         this.operator = operator;
@@ -131,7 +131,7 @@ public class SetOperationQuerySpecification<T> extends CustomQuerySpecification<
     @Override
     public SelectQueryPlan<T> createSelectPlan(int firstResult, int maxResults) {
         final String sql = getSql();
-        return new CustomSelectQueryPlan<>(extendedQuerySupport, serviceProvider, baseQuery, participatingQueries, sql, firstResult, maxResults, queryPlanCacheEnabled);
+        return new CustomSelectQueryPlan<>(extendedQuerySupport, serviceProvider, baseQuery, baseQuery, participatingQueries, sql, firstResult, maxResults, queryPlanCacheEnabled);
     }
 
     private void bindListParameters(Query q) {